+++ crates/tuirealm/src/core/mod.rs
@@ -2,18 +2,14 @@
 
 pub mod application;
 pub mod command;
-mod component;
+pub mod component;
 pub mod event;
 pub mod injector;
 pub mod props;
-mod state;
+pub mod state;
 pub mod subscription;
-mod view;
+pub mod view;
 
-// -- export
-pub use component::{Component, MockComponent};
-pub use state::{State, StateValue};
 // -- internal
 pub(crate) use subscription::Subscription;
 pub(crate) use view::WrappedComponent;
-pub use view::{View, ViewError};
